reconfigured project structure AGAIN - back to client/src and server/src
