--- config/webpack.server.dev.js
@@ -1,16 +1,0 @@
-const join = requrie('path').join;
-const readdirSync = require('fs-extra').readdirSync;
-
-const common = require('./webpack.server.common');
-
-const externals = {};
-
-readdirSync('node_modules')
-  .filter(moduleName => moduleName !== '.bin')
-  .forEach(moduleName => externals[moduleName] = `commonjs ${moduleName}`);
-
-const config = Object.assign({}, common, {
-  externals,
-});
-
-module.exports = config;
